fix checking budgets set for provider
