+++ litellm/proxy/proxy_config.yaml
@@ -12,11 +12,3 @@ model_list:
       max_budget: 10
       budget_duration: 20d
 
-litellm_settings:
-  default_internal_user_params:    # Default Params used when a new user signs in Via SSO
-    user_role: "internal_user"     # one of "internal_user", "internal_user_viewer", "proxy_admin", "proxy_admin_viewer". New SSO users not in litellm will be created as this user
-    max_budget: 100                # Optional[float], optional): $100 budget for a new SSO sign in user
-    budget_duration: 30d           # Optional[str], optional): 30 days budget_duration for a new SSO sign in user
-    models: ["gpt-3.5-turbo"]      # Optional[List[str]], optional): models to be used by a new SSO sign in user
-
-
