+++ tests/local_testing/test_router_provider_budgets.py
@@ -14,8 +14,8 @@ from litellm import Router
 from litellm.router_strategy.provider_budgets import RouterBudgetLimiting
 from litellm.types.router import (
     RoutingStrategy,
-    ProviderBudgetConfigType,
-    ProviderBudgetInfo,
+    GenericBudgetConfigType,
+    GenericBudgetInfo,
 )
 from litellm.caching.caching import DualCache, RedisCache
 import logging
@@ -59,9 +59,9 @@ async def test_provider_budgets_e2e_test():
     """
     cleanup_redis()
     # Modify for test
-    provider_budget_config: ProviderBudgetConfigType = {
-        "openai": ProviderBudgetInfo(time_period="1d", budget_limit=0.000000000001),
-        "azure": ProviderBudgetInfo(time_period="1d", budget_limit=100),
+    provider_budget_config: GenericBudgetConfigType = {
+        "openai": GenericBudgetInfo(time_period="1d", budget_limit=0.000000000001),
+        "azure": GenericBudgetInfo(time_period="1d", budget_limit=100),
     }
 
     router = Router(
@@ -208,8 +208,8 @@ async def test_get_budget_config_for_provider():
     """
     cleanup_redis()
     config = {
-        "openai": ProviderBudgetInfo(time_period="1d", budget_limit=100),
-        "anthropic": ProviderBudgetInfo(time_period="7d", budget_limit=500),
+        "openai": GenericBudgetInfo(time_period="1d", budget_limit=100),
+        "anthropic": GenericBudgetInfo(time_period="7d", budget_limit=500),
     }
 
     provider_budget = RouterBudgetLimiting(
@@ -247,15 +247,15 @@ async def test_prometheus_metric_tracking():
     provider_budget = RouterBudgetLimiting(
         router_cache=DualCache(),
         provider_budget_config={
-            "openai": ProviderBudgetInfo(time_period="1d", budget_limit=100)
+            "openai": GenericBudgetInfo(time_period="1d", budget_limit=100)
         },
     )
 
     litellm._async_success_callback = [mock_prometheus]
 
-    provider_budget_config: ProviderBudgetConfigType = {
-        "openai": ProviderBudgetInfo(time_period="1d", budget_limit=0.000000000001),
-        "azure": ProviderBudgetInfo(time_period="1d", budget_limit=100),
+    provider_budget_config: GenericBudgetConfigType = {
+        "openai": GenericBudgetInfo(time_period="1d", budget_limit=0.000000000001),
+        "azure": GenericBudgetInfo(time_period="1d", budget_limit=100),
     }
 
     router = Router(
@@ -437,8 +437,8 @@ async def test_sync_in_memory_spend_with_redis():
     """
     cleanup_redis()
     provider_budget_config = {
-        "openai": ProviderBudgetInfo(time_period="1d", budget_limit=100),
-        "anthropic": ProviderBudgetInfo(time_period="1d", budget_limit=200),
+        "openai": GenericBudgetInfo(time_period="1d", budget_limit=100),
+        "anthropic": GenericBudgetInfo(time_period="1d", budget_limit=200),
     }
 
     provider_budget = RouterBudgetLimiting(
@@ -494,7 +494,7 @@ async def test_get_current_provider_spend():
     provider_budget = RouterBudgetLimiting(
         router_cache=DualCache(),
         provider_budget_config={
-            "openai": ProviderBudgetInfo(time_period="1d", budget_limit=100),
+            "openai": GenericBudgetInfo(time_period="1d", budget_limit=100),
         },
     )
 
@@ -535,8 +535,8 @@ async def test_get_current_provider_budget_reset_at():
             )
         ),
         provider_budget_config={
-            "openai": ProviderBudgetInfo(time_period="1d", budget_limit=100),
-            "vertex_ai": ProviderBudgetInfo(time_period="1h", budget_limit=100),
+            "openai": GenericBudgetInfo(time_period="1d", budget_limit=100),
+            "vertex_ai": GenericBudgetInfo(time_period="1h", budget_limit=100),
         },
     )
 
@@ -565,3 +565,61 @@ async def test_get_current_provider_budget_reset_at():
     # Allow for small time differences (within 5 seconds)
     time_difference = abs((reset_time - expected_time).total_seconds())
     assert time_difference < 5
+
+
+@pytest.mark.asyncio
+async def test_deployment_budget_limits_e2e_test():
+    """
+    Expected behavior:
+    - First request forced to openai/gpt-4o
+    - Hit budget limit for openai/gpt-4o
+    - Next 3 requests all go to openai/gpt-4o-mini
+
+    """
+    litellm.set_verbose = True
+    cleanup_redis()
+    # Modify for test
+
+    router = Router(
+        model_list=[
+            {
+                "model_name": "gpt-4o",  # openai model name
+                "litellm_params": {  # params for litellm completion/embedding call
+                    "model": "openai/gpt-4o",
+                    "api_key": os.getenv("OPENAI_API_KEY"),
+                    "max_budget": 0.000000000001,
+                    "budget_duration": "1d",
+                },
+                "model_info": {"id": "openai-gpt-4o"},
+            },
+            {
+                "model_name": "gpt-4o",  # openai model name
+                "litellm_params": {
+                    "model": "openai/gpt-4o-mini",
+                    "api_key": os.getenv("OPENAI_API_KEY"),
+                    "max_budget": 10,
+                    "budget_duration": "20d",
+                },
+                "model_info": {"id": "openai-gpt-4o-mini"},
+            },
+        ],
+    )
+
+    response = await router.acompletion(
+        messages=[{"role": "user", "content": "Hello, how are you?"}],
+        model="openai-gpt-4o",
+    )
+    print(response)
+
+    await asyncio.sleep(2.5)
+
+    for _ in range(3):
+        response = await router.acompletion(
+            messages=[{"role": "user", "content": "Hello, how are you?"}],
+            model="gpt-4o",
+        )
+        print(response)
+        await asyncio.sleep(1)
+
+        print("response.hidden_params", response._hidden_params)
+        assert response._hidden_params.get("model_id") == "openai-gpt-4o-mini"
